+++ src/apscheduler/serializers/pickle.py
@@ -4,6 +4,7 @@ from pickle import dumps, loads
 
 import attrs
 
+from .. import DeserializationError, SerializationError
 from ..abc import Serializer
 
 
@@ -24,7 +25,13 @@ class PickleSerializer(Serializer):
     protocol: int = 4
 
     def serialize(self, obj: object) -> bytes:
-        return dumps(obj, self.protocol)
+        try:
+            return dumps(obj, self.protocol)
+        except Exception as exc:
+            raise SerializationError from exc
 
     def deserialize(self, serialized: bytes):
-        return loads(serialized)
+        try:
+            return loads(serialized)
+        except Exception as exc:
+            raise DeserializationError from exc