+++ tests/test_serializers.py
@@ -1,11 +1,19 @@
 from __future__ import annotations
 
 from datetime import datetime, timezone
+from typing import NoReturn
 from uuid import uuid4
 
 import pytest
 
-from apscheduler import Event, JobAdded, JobOutcome, JobReleased
+from apscheduler import (
+    DeserializationError,
+    Event,
+    JobAdded,
+    JobOutcome,
+    JobReleased,
+    SerializationError,
+)
 from apscheduler.abc import Serializer
 
 
@@ -38,3 +46,29 @@ def test_serialize_event(event: Event, serializer: Serializer) -> None:
     payload = serializer.serialize(event.marshal())
     deserialized = type(event).unmarshal(serializer.deserialize(payload))
     assert deserialized == event
+
+
+def test_serialization_error(serializer: Serializer) -> None:
+    class Unserializable:
+        def __getstate__(self) -> NoReturn:
+            raise ValueError("cannot be serialized")
+
+    # An open file cannot be serialized
+    with pytest.raises(SerializationError) as exc:
+        serializer.serialize(Unserializable())
+
+    assert isinstance(exc.value.__cause__, ValueError)
+
+
+@pytest.mark.parametrize(
+    "payload",
+    [
+        pytest.param(b"", id="empty"),
+        pytest.param(b"\x61\x98", id="invalid"),
+    ],
+)
+def test_deserialization_error(payload: bytes, serializer: Serializer) -> None:
+    with pytest.raises(DeserializationError) as exc:
+        serializer.deserialize(payload)
+
+    assert exc.value.__cause__ is not None